Add lsof (used by the gdbdebugger plugin) as a dependency
The gdbdebugger plugin implementation uses lsof and greps for gdbserver to
see if it has started, and keeps looping until it does.

On a standard arch linux box, lsof is not installed by default, and this
causes the plugin to assume gdbserver hasn't spun up - it ends up looping
endlessly. Takes some time and debugging effort to figure out the actual problem.